refactor: replace with new agent implementation
