+++ freeplay/trajectory_runner.py
@@ -5,13 +5,21 @@ from dataclasses import dataclass
 import multiprocessing
 from dotenv import load_dotenv
 
-from agents.agent_abc import AgentABC
 from freeplay.evaluator import SimpleFactorioEvaluator
 from models.conversation import Conversation
 from models.message import Message
 from models.program import Program
 from instance import FactorioInstance
-from freeplay.basic_agent import BasicAgent
+from trainer.agent import BasicAgent
+from trainer.definitions import (
+    Step,
+    Execution,
+    ParsedGameState,
+    AgentOutput,
+    Evaluation,
+    DataPoint,
+    create_data_point,
+)
 
 from namespace import FactorioNamespace
 
@@ -50,6 +58,37 @@ def format_inventory(inventory: dict) -> str:
     return f"{inventory}, {max(80 - slot, 0)} slots remaining"
 
 
+def execution_history_from_conversation(conversation: Conversation) -> List[Execution]:
+    """Convert conversation to execution history"""
+    history = []
+    for i in range(1, len(conversation.messages), 2):
+        code = conversation.messages[i].content
+        response = conversation.messages[i + 1].content
+
+        history.append(
+            Execution(
+                step=None,
+                game_state=None,
+                agent_output=AgentOutput(
+                    input_messages=[],
+                    raw_response="",
+                    thinking="",
+                    code=code,
+                ),
+                evaluation=Evaluation(
+                    game_state=None,
+                    response=response,
+                    reward=0,
+                    achievements={},
+                    flows=None,
+                    ticks=0,
+                ),
+            )
+        )
+
+    return history
+
+
 class TrajectoryRunner:
     """Handles program generation and evaluation for a single trajectory"""
 
@@ -74,20 +113,21 @@ class TrajectoryRunner:
 
     async def _generate_program(
         self,
+        step: Step,
+        game_state: ParsedGameState,
         conversation: Conversation,
         response: Response,
-        namespace: FactorioNamespace,
-        entities: str,
-        inventory: str,
         meta={},
     ) -> Program:
         conversation = copy.deepcopy(conversation)
         try:
-            policy = await self.agent.step(
-                conversation, response, namespace, entities, inventory
+            agent_output = await self.agent.run(
+                step=step,
+                game_state=game_state,
+                execution_history=execution_history_from_conversation(conversation),
             )
 
-            if not policy:
+            if not agent_output:
                 raise Exception("Policy not valid Python. Skipping.")
 
             try:
@@ -96,24 +136,24 @@ class TrajectoryRunner:
                 messages = conversation.dict()["messages"]
 
             program = Program(
-                thinking=policy.thinking,
-                code=policy.code,
+                thinking=agent_output.thinking,
+                code=agent_output.code,
                 conversation=conversation,
                 response=response.response if response else None,
-                token_usage=policy.meta.total_tokens,
-                completion_token_usage=policy.meta.output_tokens,
-                prompt_token_usage=policy.meta.input_tokens,
+                # token_usage=policy.meta.total_tokens,
+                # completion_token_usage=policy.meta.output_tokens,
+                # prompt_token_usage=policy.meta.input_tokens,
                 version=self.config.version,
-                model=self.agent.model,
+                model=self.config.model,
                 version_description=self.config.version_description,
-                meta={"model": self.agent.model, "process_id": self.process_id},
+                meta={"model": self.config.model, "process_id": self.process_id},
                 depth=len(messages) - 2,
             )
 
             if meta:
                 program.meta.update(meta)
 
-            return program
+            return agent_output, program
 
         except Exception as e:
             print(f"Program generation failed: {str(e)}")
@@ -139,7 +179,6 @@ class TrajectoryRunner:
             ) = await self.db.get_resume_state(
                 resume_version=self.config.version, process_id=self.process_id
             )
-            self.agent.conversation = current_conversation
 
             if current_state:
                 instance = self.evaluator.instance
@@ -156,7 +195,7 @@ class TrajectoryRunner:
 
         # New game
         if not current_state:
-            current_state = self.agent.task.starting_game_state
+            current_state = self.config.task.starting_game_state
             depth = 0
             instance = self.evaluator.instance
             instance.reset(current_state)
@@ -176,7 +215,6 @@ class TrajectoryRunner:
                     ),
                 ]
             )
-            self.agent.conversation = current_conversation
             parent_id = None
 
         # with open("entities.txt", "w") as f:
@@ -226,12 +264,6 @@ class TrajectoryRunner:
 
                 time.sleep(60)
 
-            await self.agent.start_iteration(
-                iteration=iteration,
-                instruction=instruction,
-                previous_iteration_summary=previous_iteration_summary,
-            )
-
             update_spreadsheet_cell(
                 os.getenv("SPREADSHEET_ID"),
                 "System!B1",
@@ -266,17 +298,24 @@ class TrajectoryRunner:
                     )
 
                     print("generation starting...")
-                    program = await self._generate_program(
-                        current_conversation,
-                        last_response,
-                        self.evaluator.instance.namespace,
-                        entities=current_entities,
-                        inventory=current_inventory,
+                    agent_output, program = await self._generate_program(
+                        step=Step(
+                            number=depth + 1,
+                            instruction=instruction,
+                            iteration_number=iteration,
+                            in_iteration_number=step + 1,
+                        ),
+                        game_state=ParsedGameState(
+                            raw=current_state,
+                            entities=current_entities,
+                        ),
+                        conversation=current_conversation,
+                        response=last_response,
                     )
 
                     print(
                         f"Generated program {multiprocessing.current_process().name} - "
-                        f"Model: {self.agent.model} - "
+                        f"Model: {self.config.model} - "
                         f"Step {iteration}-{step + 1}"
                     )
 
@@ -311,7 +350,7 @@ class TrajectoryRunner:
                         evaluated_program,
                         task_verification_response,
                     ) = await self.evaluator.evaluate(
-                        program, current_state, iteration, instruction, self.agent.task
+                        program, current_state, iteration, instruction, self.config.task
                     )
                     print(program.code + "\n" + "=" * 50)
                     print(
@@ -327,7 +366,7 @@ class TrajectoryRunner:
                     )
                     print(
                         f"Evaluated program {multiprocessing.current_process().name} - "
-                        f"Model: {self.agent.model} - "
+                        f"Model: {self.config.model} - "
                         f"Step {iteration}-{step + 1}"
                     )
 
@@ -335,8 +374,7 @@ class TrajectoryRunner:
                         continue
 
                     program = evaluated_program
-                    self.agent.conversation = program.conversation
-                    program.meta["task_key"] = self.agent.task.task_key
+                    program.meta["task_key"] = self.config.task.task_key
                     last_response = Response(
                         code=program.code,
                         created_at=program.created_at,
@@ -353,7 +391,7 @@ class TrajectoryRunner:
                     saved_program = await self.db.create_program(program)
                     print(
                         f"Saved program {multiprocessing.current_process().name} - "
-                        f"Model: {self.agent.model} - "
+                        f"Model: {self.config.model} - "
                         f"Step {iteration}-{step + 1}"
                     )
 
@@ -364,6 +402,40 @@ class TrajectoryRunner:
                             os.getenv("SPREADSHEET_ID"),
                             f"Steps!J{step_row_number}",
                             program.response,
+                        )
+
+                    with open("execution.json", "w") as f:
+                        json.dump(
+                            create_data_point(
+                                collection_id="test",
+                                agent_name="BasicAgent",
+                                agent_version="1",
+                                execution=Execution(
+                                    step=Step(
+                                        number=depth + 1,
+                                        instruction=instruction,
+                                        iteration_number=iteration,
+                                        in_iteration_number=step + 1,
+                                    ),
+                                    game_state=ParsedGameState(
+                                        raw=current_state,
+                                        entities=current_entities,
+                                    ),
+                                    agent_output=agent_output,
+                                    evaluation=Evaluation(
+                                        game_state=program.state,
+                                        response=program.response,
+                                        reward=program.value,
+                                        achievements=program.achievements,
+                                        flows=program.flows,
+                                        ticks=program.ticks,
+                                    ),
+                                ),
+                                execution_history=execution_history_from_conversation(
+                                    program.conversation
+                                ),
+                            ).to_json(),
+                            f,
                         )
 
                     # Update state for next iteration
@@ -396,7 +468,7 @@ class TrajectoryRunner:
             elapsed_str = f"{int(elapsed // 3600):02d}:{int((elapsed % 3600) // 60):02d}:{int(elapsed % 60):02d}"
             print(
                 f"\033[92m Process {multiprocessing.current_process().name} - "
-                f"Model: {self.agent.model} - "
+                f"Model: {self.config.model} - "
                 f"Itertion {iteration} - "
                 f"Value: {program.value:.2f} - "
                 f"Elapsed: {elapsed_str} - "
@@ -445,9 +517,7 @@ async def run_trajectory(process_id: int, config: PlayConfig):
         instance=instance, value_accrual_time=1, error_penalty=0
     )
 
-    agent = BasicAgent(
-        model=config.model, system_prompt=system_prompt, task=config.task
-    )
+    agent = BasicAgent(model=config.model, system_prompt=system_prompt)
 
     # setup the instance
     task = config.task
